LayoutTests/fast/forms/select/base/select-pagedown-pageup-vertical-expected.txt

Tests Page Down and Page Up keyboard navigation in a base-select picker with vertical writing modes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



vertical-lr:

PASS focusedOptionNumber() is 50
PASS isFocusedOptionVisible() is true
PASS focusedOptionNumber() is 1
PASS focusedOptionNumber() is >= 2
PASS focusedOptionNumber() is 100
PASS focusedOptionNumber() < 100 is true
PASS focusedOptionNumber() is 100
PASS focusedOptionNumber() is 1
PASS focusedOptionNumber() === pageDownTarget is true
PASS focusedOptionNumber() === pageUpTarget is true

vertical-rl:

PASS focusedOptionNumber() is 50
PASS isFocusedOptionVisible() is true
PASS focusedOptionNumber() is 1
PASS focusedOptionNumber() is >= 2
PASS focusedOptionNumber() is 100
PASS focusedOptionNumber() < 100 is true
PASS focusedOptionNumber() is 100
PASS focusedOptionNumber() is 1
PASS focusedOptionNumber() === pageDownTarget is true
PASS focusedOptionNumber() === pageUpTarget is true
PASS successfullyParsed is true

TEST COMPLETE
